feat(wip): UdonChips対応版を継承classで作成
