+++ Assets/SerializedUdonPrograms/f4ce24999762a9b48b2b2abe781896c9.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: af488b68e1e5ca043a34b3cf92b27ea5
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
