+++ Assets/UdonKnucklebonesSupportUdonChips.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 5730f95a092738e4abd01aa280bc6a09
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
