+++ Assets/UdonKnucklebonesSupportUdonChips/Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 8265f9e3b0e88f54daf484fcc7c27dbb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
