+++ Assets/UdonKnucklebonesSupportUdonChips/Udon.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: e0aff1c68b060164b8b45c8802f5c7a2
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
