+++ Assets/UdonKnucklebonesSupportUdonChips/Udon/UdonKnucklebonesSupportUdonChips.asset.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: f4ce24999762a9b48b2b2abe781896c9
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 11400000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
