+++ Assets/UdonKnucklebonesSupportUdonChips/Udon/UdonKnucklebonesSupportUdonChips.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: eeaaf91993f0a5047a57f20fe9610550
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
