+++ Assets/UdonKnucklebonesSupportUdonChips/Udon/Wipeseals.UdonKnucklebonesSupportUdonChips.asmref
@@ -1,0 +1,3 @@
+{
+    "reference": "GUID:5786ab134f4cf43418a609d9963d81af"
+}
