+++ Assets/UdonKnucklebonesSupportUdonChips/Udon/Wipeseals.UdonKnucklebonesSupportUdonChips.asmref.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4659690c506a9a7428691b1a11b60e67
+AssemblyDefinitionReferenceImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
