+++ Packages/me.wipeseals.udon-knucklebones/Runtime/Udon/UdonKnucklebones.cs
@@ -1,17 +1,10 @@
-﻿// UdonChips対応を有効にする場合は定義する
-// #define UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
-
-using UdonSharp;
+﻿using UdonSharp;
 using UnityEngine;
 using UnityEngine.UI;
 using VRC.SDKBase;
 using VRC.Udon;
 using TMPro;
 using System;
-
-#if UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
-using UCS;
-#endif // UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
 
 namespace Wipeseals
 {
@@ -880,78 +873,22 @@ namespace Wipeseals
         }
 
         #endregion
-        #region UdonChips Utility
-#if UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
+        #region UdonChips Event (Empty)
         /// <summary>
         /// 所持金を取得し同期変数に設定
+        /// 本家UdonKnucklebonesではUdonChipsへの参照を持っていないので何もしない
         /// </summary>
-        void UpdateCurrentUdonChips()
+        public virtual void OnUpdateCurrentUdonChips()
         {
-            Log(ErrorLevel.Info, $"{nameof(UpdateCurrentUdonChips)}");
-
-            var money = GameObject.Find("UdonChips").GetComponent<UdonChips>().money;
-            if (IsMyselfPlayer1)
-            {
-                Player1UdonChips = money;
-            }
-            else if (IsMyselfPlayer2)
-            {
-                Player2UdonChips = money;
-            }
         }
 
         /// <summary>
         /// 勝敗の金額を反映。ローカル処理
+        /// 本家UdonKnucklebonesではUdonChipsへの参照を持っていないので何もしない
         /// </summary>
-        void ApplyUdonChips()
+        public virtual void OnApplyUdonChips()
         {
-            // UdonChips自体はLocalなのでOwnerでなくても問題ない
-
-            // 取引金額を計算
-            var player1Score = GetDiceArrayBits(PLAYER1).GetTotalScore();
-            var player2Score = GetDiceArrayBits(PLAYER2).GetTotalScore();
-            var scoreDiff = (player1Score > player2Score) ? player1Score - player2Score : player2Score - player1Score;
-
-            var ratio = ((Player1Type == (int)PlayerType.CPU || Player2Type == (int)PlayerType.CPU) ? UdonChipsCpuRate : UdonChipsPlayerRate);
-            var applyMoney = scoreDiff * ratio;
-
-            // 負けたPlayerが支払えないケースでは残金全てに設定。CPUの場合は全額のまま
-            if (CurrentGameJudge == (int)GameJudge.Player1Win && Player1Type == (int)PlayerType.Human && Player2UdonChips < applyMoney)
-            {
-                applyMoney = Player2UdonChips;
-            }
-            else if (CurrentGameJudge == (int)GameJudge.Player2Win && Player2Type == (int)PlayerType.Human && Player1UdonChips < applyMoney)
-            {
-                applyMoney = Player1UdonChips;
-            }
-
-            // 取引。それぞれのローカルでmoneyを更新
-            if (IsMyselfPlayer1)
-            {
-                if (CurrentGameJudge == (int)GameJudge.Player1Win)
-                {
-                    GameObject.Find("UdonChips").GetComponent<UdonChips>().money += applyMoney;
-                }
-                else if (CurrentGameJudge == (int)GameJudge.Player2Win)
-                {
-                    GameObject.Find("UdonChips").GetComponent<UdonChips>().money -= applyMoney; // 事前に支払えないケースの対応は済んでいるので、ここではそのまま減算
-                }
-            }
-            else if (IsMyselfPlayer2)
-            {
-                if (CurrentGameJudge == (int)GameJudge.Player2Win)
-                {
-                    GameObject.Find("UdonChips").GetComponent<UdonChips>().money += applyMoney;
-                }
-                else if (CurrentGameJudge == (int)GameJudge.Player1Win)
-                {
-                    GameObject.Find("UdonChips").GetComponent<UdonChips>().money -= applyMoney; // 事前に支払えないケースの対応は済んでいるので、ここではそのまま減算
-                }
-            }
-
-            Log(ErrorLevel.Info, $"Player1Score={player1Score} Player2Score={player2Score} ScoreDiff={scoreDiff} ApplyMoney={applyMoney}");
         }
-#endif // UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
         #endregion
         #region UI Utility
 
@@ -1247,7 +1184,7 @@ namespace Wipeseals
         /// Prints a message to the console
         /// </summary>
         /// <param name="msg"></param>
-        void Log(ErrorLevel level, string msg)
+        internal void Log(ErrorLevel level, string msg)
         {
             switch (level)
             {
@@ -1645,11 +1582,9 @@ namespace Wipeseals
             }
 
 
-#if UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
             // 毎ターンごとに各PlayerがOwnerを持ち、変数更新を行えるのでUdonChips最新値を取得しておく
             // 差額計算に近いタイミングではあるが、反映までの僅かな間に変更されるケースは諦める (一応Underflowしない対策入れた)
-            UpdateCurrentUdonChips();
-#endif // UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
+            OnUpdateCurrentUdonChips();
 
             // 左詰めの処理していないのでここでやる（PutDice時の消えるアニメーション流したいため)
             SetDiceArrayBits(PLAYER1, GetDiceArrayBits(PLAYER1).LeftJustify());
@@ -2337,11 +2272,8 @@ namespace Wipeseals
                 return;
             }
 
-#if UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
-            // ゲーム終了に合わせて残金精算。条件付きコンパイルのため非サポートの場合は何もしない
-            ApplyUdonChips();
-#endif // UDON_KNUCKLEBONES_SUPPORT_UDONCHIPS
-
+            // ゲーム終了に合わせて残金精算
+            OnApplyUdonChips();
         }
         #endregion
     }
