+++ Packages/me.wipeseals.udon-knucklebones/Runtime/Wipeseals.UdonKnucklebones.asmdef
@@ -12,7 +12,7 @@
     "allowUnsafeCode": false,
     "overrideReferences": false,
     "precompiledReferences": [],
-    "autoReferenced": false,
+    "autoReferenced": true,
     "defineConstraints": [],
     "versionDefines": [],
     "noEngineReferences": false
